app/requirements.txt
Flask
mysql-connector
requests
json